Java Version
For this laboratory, Java seemed to offer more control over the tables used to describe the NFA or the DFA, because it allows us to specify the data type for every single piece of information and the structure that stores it. A Python version might be on its way soon.